Cleanup "$FlowFixMe (T62907961)" No errors were added from store mutations. Errors were added because of incorrect graphql type imports 431 errors

+++ app/fbcnms-projects/inventory/app/components/configure/LocationTypeItem.js
@@ -125,18 +125,17 @@ class LocationTypeItem extends React.Component<Props> {
             },
           },
           store => {
-            // $FlowFixMe (T62907961) Relay flow types
             const rootQuery = store.getRoot();
             const locationTypes = ConnectionHandler.getConnection(
               rootQuery,
               'Catalog_locationTypes',
             );
-            ConnectionHandler.deleteNode(
-              // $FlowFixMe (T62907961) Relay flow types
-              locationTypes,
-              this.props.locationType.id,
-            );
-            // $FlowFixMe (T62907961) Relay flow types
+            if (locationTypes != null) {
+              ConnectionHandler.deleteNode(
+                locationTypes,
+                this.props.locationType.id,
+              );
+            }
             store.delete(this.props.locationType.id);
           },
         );
