+++ app/fbcnms-projects/inventory/app/components/configure/ServiceTypeItem.js
@@ -108,18 +108,17 @@ class ServiceTypeItem extends React.Component<Props> {
             },
           },
           store => {
-            // $FlowFixMe (T62907961) Relay flow types
             const rootQuery = store.getRoot();
             const serviceTypes = ConnectionHandler.getConnection(
               rootQuery,
               'ServiceTypes_serviceTypes',
             );
-            ConnectionHandler.deleteNode(
-              // $FlowFixMe (T62907961) Relay flow types
-              serviceTypes,
-              this.props.serviceType.id,
-            );
-            // $FlowFixMe (T62907961) Relay flow types
+            if (serviceTypes != null) {
+              ConnectionHandler.deleteNode(
+                serviceTypes,
+                this.props.serviceType.id,
+              );
+            }
             store.delete(this.props.serviceType.id);
           },
         );
